Flutter
Getting started with Flutter Development" is an online non-credit project authorized by Google Cloud and offered through Coursera. This project aims to help developers acquire the necessary skills to start developing applications with Flutter, an open-source framework for building user interfaces for mobile, web, and desktop applications. The course includes multiple modules covering the fundamental concepts of Flutter and provides practical tutorials and projects for learners to apply the concepts learned in real-world projects. Upon completion of the project, learners will have a solid understanding of the basics of Flutter and will be able to create effective and high-quality mobile applications.