Data base intro
"Foundations: Data, Data, Everywhere" is an online non-credit course authorized by Google and offered through Coursera. This course provides an introduction to the basics of data analytics, including data management, data visualization, and data-driven decision-making. Through hands-on exercises and projects, learners will gain practical experience in using tools and techniques to manage and analyze different types of data. By the end of the course, learners will have a solid understanding of data analytics and be able to apply their knowledge to real-world problems. Overall, this course is a great starting point for learners who are interested in data analytics and want to develop practical skills in this field.