Material Components for Flutter Basics
"Material Components for Flutter Basics" is an online non-credit project authorized by Google Cloud and offered through Coursera. This project provides an introduction to building user interfaces in Flutter using the Material Design system. Throughout the project, learners will gain practical experience in implementing different types of Material Design components and managing app state and navigation using Flutter's built-in widgets. By the end of the project, learners will have a solid understanding of the Material Design system and how to use it to create visually appealing user interfaces in Flutter. This project is a great choice for learners who want to develop practical skills in building user interfaces with Flutter and using the Material Design system.