Fundamentals of Digital Marketing
"Fundamentals of Digital Marketing" and it is offered by Google Digital Garage. The certification program covers the basics of digital marketing, including various channels such as social media, search engine optimization (SEO), email marketing, and advertising. Learners engage in practical exercises and projects throughout the course to gain hands-on experience in developing digital marketing strategies. Upon completion of the program, learners receive a certification from Google Digital Garage, which can help enhance their credibility and career prospects in the digital marketing industry. Overall, this certification is a great choice for learners who are interested in digital marketing and want to develop practical skills in this field.