Programming in C
"Programming in C" and it is offered by Sololearn. The course covers the basics of programming in C, including its syntax, control structures, data types, functions, and pointers. Throughout the course, learners engage in practical exercises and projects to gain hands-on experience in developing software applications in C. Upon completion of the course, learners receive a certification from Sololearn, which can help enhance their credibility and career prospects in the software development industry. Overall, this certification is a great choice for learners who want to develop practical skills in programming in C or pursue a career in software development.